Moved a couple of tooling model builder classes to live with the other builder classes.
