+++ subprojects/ide/src/main/groovy/org/gradle/plugins/ide/internal/tooling/PublicationsBuilder.java
@@ -23,7 +23,7 @@ import org.gradle.api.internal.artifacts.ivyservice.projectmodule.ProjectPublica
 import org.gradle.tooling.internal.gradle.DefaultGradleModuleVersion;
 import org.gradle.tooling.internal.gradle.DefaultGradlePublication;
 import org.gradle.tooling.internal.gradle.DefaultProjectPublications;
-import org.gradle.tooling.model.internal.ProjectSensitiveToolingModelBuilder;
+import org.gradle.tooling.provider.model.internal.ProjectSensitiveToolingModelBuilder;
 
 import java.util.List;
 import java.util.Set;
