+++ subprojects/tooling-api-builders/src/main/java/org/gradle/tooling/internal/provider/runner/BuildModelsActionRunner.java
@@ -24,7 +24,7 @@ import org.gradle.internal.invocation.BuildActionRunner;
 import org.gradle.internal.invocation.BuildController;
 import org.gradle.tooling.internal.protocol.InternalUnsupportedModelException;
 import org.gradle.tooling.internal.provider.BuildModelAction;
-import org.gradle.tooling.model.internal.ProjectToolingModelBuilder;
+import org.gradle.tooling.provider.model.internal.ProjectToolingModelBuilder;
 import org.gradle.tooling.provider.model.ToolingModelBuilder;
 import org.gradle.tooling.provider.model.ToolingModelBuilderRegistry;
 import org.gradle.tooling.provider.model.UnknownModelException;
